Add VertxPinotClientTransport and test for same
BrokerResponse is a package-private class in org.apache.pinot.client
package. However to implement the PinotClientTransport class, we need
to return the BrokerResponse class from our implementation.

To workaround this, we require a dirty hack to create an
org.apache.pinot.client in our project and put the VertxPinotClientTransport
class in it.